Remove dependency on adapter.py from events.py for Plone < 5.2
